refactor(agents): tighten session lock modules
