--- src/agents/embedded-agent-runner/run/attempt.session-lock.error.ts
@@ -1,6 +1,0 @@
-export class EmbeddedAttemptSessionTakeoverError extends Error {
-  constructor(sessionFile: string) {
-    super(`session file changed while embedded prompt lock was released: ${sessionFile}`);
-    this.name = "EmbeddedAttemptSessionTakeoverError";
-  }
-}
