+++ src/agents/embedded-agent-runner/run/attempt.session-lock.ts
@@ -5,7 +5,7 @@ export {
   createEmbeddedAttemptSessionLockController,
   type EmbeddedAttemptSessionLockController,
 } from "./attempt.session-lock.controller.js";
-export { EmbeddedAttemptSessionTakeoverError } from "./attempt.session-lock.error.js";
+export { EmbeddedAttemptSessionTakeoverError } from "./attempt.session-lock.fence-controller.js";
 export {
   acquireEmbeddedAttemptSessionFileOwner,
   type EmbeddedAttemptSessionFileOwner,
