+++ src/agents/embedded-agent-runner/run/attempt.session-lock.write-scope.ts
@@ -1,5 +1,3 @@
-import type { OwnedSessionTranscriptPublishedEntry } from "../../../config/sessions/transcript-write-context.js";
-
 type PhysicalWriteLockScope = {
   active: boolean;
   completion: Promise<void>;
